ch11/k8s-ai/pyproject.toml
[project]
name = "k8s-ai"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "openai",
    "sh",
    "a2a-sdk[http-server]",
    "uvicorn",
    "kubernetes>=29.0.0",
    "pyyaml>=6.0.1",
    "fastapi>=0.104.0",
    "structlog>=23.2.0",
]

[project.scripts]
k8s-ai-cli = "k8s_ai.cli.main:main"
k8s-ai-server = "k8s_ai.server.main:main"

[tool.uv]
package = true